+++ app/src/main/java/com/github/scribeWizTeam/scribewiz/models/BadgeModel.kt
@@ -1,11 +1,14 @@
 package com.github.scribeWizTeam.scribewiz.models
 
+import com.google.firebase.firestore.DocumentReference
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.firestore.ktx.toObject
 import com.google.firebase.ktx.Firebase
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.tasks.await
+import java.text.DateFormat
+import java.util.*
 
 data class BadgeModel(override var id: String? = "",
                       var badgeName: String? = "",
@@ -40,6 +43,36 @@ data class BadgeModel(override var id: String? = "",
                 job.join()
             }
             return badges
+        }
+
+        fun addBadgeToUser(user: UserModel){
+            val time = Calendar.getInstance().time
+            val formatter = DateFormat.getDateTimeInstance()
+            val badgeData = BadgeModel(
+                "123",
+                "Test Badge",
+                "abc",
+                formatter.format(time),
+                ""
+            )
+
+            val userDoc : DocumentReference =
+                Firebase.firestore
+                .collection("Users")
+                .document(user.id!!)
+
+            // Add badge in badge collection
+            userDoc.collection("Badges")
+                .document(badgeData.id!!)
+                .set(badgeData)
+
+            // Add badge in badge list
+            user.badges!!.add(badgeData.id.toString())
+
+            // Update user and badge in DB
+            user.updateInDB()
+            badgeData.updateInDB()
+
         }
     }
